refactor(davinci): keep the P0-4 budget gate under the source-length ceiling
The 350-line source ceiling (tests/tooling/source-file-lengths.test.ts) failed
on all three P0-4 files. budgets.toml now writes each uniform four-field bench
entry as a one-line inline table, which teaches tools/davinci/toml-lite.mjs the
single-line inline-table shape and keeps the registry compact as benches are
added. bench-compare.mjs keeps the CLI and moves the registry loader, report
loader, and verdict rows into tools/davinci/lib/bench-*.mjs, and the compare
gate's stdout oracles move to tests/tooling/davinci-bench-compare.test.ts.

No gate semantics change: the same 85 ids, tolerances, verdict rows, and exit
codes.
