+++ tests/tooling/davinci-bench-compare.test.ts
@@ -1,0 +1,210 @@
+// The davinci bench budget compare gate (plan/phase-0.md P0-4).
+//
+// Exact stdout/stderr/exit oracles for tools/davinci/bench-compare.mjs over
+// the committed fixture pairs in tests/_fixtures/davinci-bench-compare/,
+// including the DAVINCI_BASELINE_REFRESH refusal path. The registry side of
+// P0-4 (budgets.toml reconciliation and the ratchet header) lives in
+// tests/tooling/davinci-budgets.test.ts.
+
+import assert from "node:assert/strict";
+import { spawnSync } from "node:child_process";
+import fs from "node:fs";
+import os from "node:os";
+import path from "node:path";
+import { test } from "node:test";
+import { fileURLToPath } from "node:url";
+
+const repoRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "../..");
+const comparePath = path.join(repoRoot, "tools", "davinci", "bench-compare.mjs");
+const fixtureRel = "tests/_fixtures/davinci-bench-compare";
+const fixtureDir = path.join(repoRoot, fixtureRel);
+
+function runCompare(
+  args: string[],
+  options: { cwd?: string; refreshEnv?: string } = {},
+): ReturnType<typeof spawnSync<string>> {
+  const env = { ...process.env };
+  delete env.DAVINCI_BASELINE_REFRESH;
+  if (options.refreshEnv != null) env.DAVINCI_BASELINE_REFRESH = options.refreshEnv;
+  return spawnSync(process.execPath, [comparePath, ...args], {
+    cwd: options.cwd ?? repoRoot,
+    encoding: "utf8",
+    env,
+  });
+}
+
+function compareArgs(scenario: string, overrides: { budgets?: string; baseline?: string } = {}) {
+  return [
+    "--budgets",
+    overrides.budgets ?? `${fixtureRel}/budgets.toml`,
+    "--baseline",
+    overrides.baseline ?? `${fixtureRel}/baseline`,
+    "--results",
+    `${fixtureRel}/${scenario}/current`,
+  ];
+}
+
+function header(scenario: string, overrides: { budgets?: string; baseline?: string } = {}) {
+  return (
+    `bench-compare: budgets=${overrides.budgets ?? `${fixtureRel}/budgets.toml`} ` +
+    `baseline=${overrides.baseline ?? `${fixtureRel}/baseline`} ` +
+    `results=${fixtureRel}/${scenario}/current\n`
+  );
+}
+
+test("bench-compare exits 0 when the current run sits inside each bench's own tolerance", () => {
+  const result = runCompare(compareArgs("within-tolerance"));
+  assert.equal(result.stderr, "");
+  assert.equal(result.status, 0, result.stdout);
+  // The +8% transform run is inside its 0.10 stage-window tolerance but would
+  // breach the 0.05 whole-routine one, so this also pins per-bench tolerance.
+  assert.equal(
+    result.stdout,
+    header("within-tolerance") +
+      "ok fixture_parse_small wall_p50 104000ns (baseline 100000ns limit 105000ns) " +
+      "allocs 42 rss 8192B\n" +
+      "ok fixture_transform_medium wall_p50 216000ns (baseline 200000ns limit 220000ns) " +
+      "allocs 1000 rss 16384B\n" +
+      "bench-compare: breaches=0 gated_ok=2 report_only=0 registered=2\n",
+  );
+});
+
+test("bench-compare exits 1 when wall p50 breaches the baseline tolerance", () => {
+  const result = runCompare(compareArgs("wall-breach"));
+  assert.equal(result.stderr, "");
+  assert.equal(result.status, 1, result.stdout);
+  assert.equal(
+    result.stdout,
+    header("wall-breach") +
+      "FAIL fixture_parse_small wall_p50 106000ns > limit 105000ns " +
+      "(baseline 100000ns + 5% tolerance)\n" +
+      "ok fixture_transform_medium wall_p50 216000ns (baseline 200000ns limit 220000ns) " +
+      "allocs 1000 rss 16384B\n" +
+      "bench-compare: breaches=1 gated_ok=1 report_only=0 registered=2\n",
+  );
+});
+
+test("bench-compare exits 1 on any allocation-count change (exact gate)", () => {
+  const result = runCompare(compareArgs("allocs-breach"));
+  assert.equal(result.stderr, "");
+  assert.equal(result.status, 1, result.stdout);
+  assert.equal(
+    result.stdout,
+    header("allocs-breach") +
+      "FAIL fixture_parse_small allocs 42 -> 43 (exact gate: allocs are deterministic)\n" +
+      "ok fixture_transform_medium wall_p50 216000ns (baseline 200000ns limit 220000ns) " +
+      "allocs 1000 rss 16384B\n" +
+      "bench-compare: breaches=1 gated_ok=1 report_only=0 registered=2\n",
+  );
+});
+
+test("bench-compare exits 1 on registry drift in either direction", () => {
+  const result = runCompare(compareArgs("drift"));
+  assert.equal(result.stderr, "");
+  assert.equal(result.status, 1, result.stdout);
+  assert.equal(
+    result.stdout,
+    header("drift") +
+      "ok fixture_parse_small wall_p50 104000ns (baseline 100000ns limit 105000ns) " +
+      "allocs 42 rss 8192B\n" +
+      "FAIL fixture_transform_medium bench disappeared " +
+      "(budgets.toml entry has no current result)\n" +
+      "FAIL fixture_unknown_new unregistered bench " +
+      "(current result has no budgets.toml [bench] entry)\n" +
+      "bench-compare: breaches=2 gated_ok=1 report_only=0 registered=2\n",
+  );
+});
+
+test("bench-compare lists missing-baseline benches as unbaselined and report-only", () => {
+  const overrides = { baseline: `${fixtureRel}/no-baseline` };
+  const result = runCompare(compareArgs("within-tolerance", overrides));
+  assert.equal(result.stderr, "");
+  assert.equal(result.status, 0, result.stdout);
+  assert.equal(
+    result.stdout,
+    header("within-tolerance", overrides) +
+      "unbaselined fixture_parse_small wall_p50 104000ns allocs 42 rss 8192B (report-only)\n" +
+      "unbaselined fixture_transform_medium wall_p50 216000ns allocs 1000 rss 16384B " +
+      "(report-only)\n" +
+      "bench-compare: breaches=0 gated_ok=0 report_only=2 registered=2\n",
+  );
+});
+
+test("bench-compare treats all-zero (unrecorded) budget entries as report-only", () => {
+  const overrides = { budgets: `${fixtureRel}/budgets-unrecorded.toml` };
+  const result = runCompare(compareArgs("within-tolerance", overrides));
+  assert.equal(result.stderr, "");
+  assert.equal(result.status, 0, result.stdout);
+  assert.equal(
+    result.stdout,
+    header("within-tolerance", overrides) +
+      "unrecorded fixture_parse_small wall_p50 104000ns allocs 42 rss 8192B " +
+      "(report-only: budgets.toml baseline not yet recorded)\n" +
+      "unrecorded fixture_transform_medium wall_p50 216000ns allocs 1000 rss 16384B " +
+      "(report-only: budgets.toml baseline not yet recorded)\n" +
+      "bench-compare: breaches=0 gated_ok=0 report_only=2 registered=2\n",
+  );
+});
+
+const refusalMessage =
+  "bench-compare: refusing --update-baseline without DAVINCI_BASELINE_REFRESH=1.\n" +
+  "The committed baseline is the reference every PR is gated against; refreshing it\n" +
+  "must be a deliberate act on the reference runner, not a side effect. Re-run with\n" +
+  "DAVINCI_BASELINE_REFRESH=1 in the environment to proceed.\n";
+
+function baselineSnapshot(dir: string): Record<string, string> {
+  const snapshot: Record<string, string> = {};
+  for (const name of fs.readdirSync(dir).sort()) {
+    snapshot[name] = fs.readFileSync(path.join(dir, name), "utf8");
+  }
+  return snapshot;
+}
+
+test("--update-baseline refuses without DAVINCI_BASELINE_REFRESH=1 and writes nothing", () => {
+  const baselineDir = path.join(fixtureDir, "baseline");
+  const before = baselineSnapshot(baselineDir);
+  for (const refreshEnv of [undefined, "0"]) {
+    const result = runCompare([...compareArgs("within-tolerance"), "--update-baseline"], {
+      refreshEnv,
+    });
+    assert.equal(result.status, 2, result.stderr);
+    assert.equal(result.stdout, "");
+    assert.equal(result.stderr, refusalMessage);
+  }
+  assert.deepEqual(baselineSnapshot(baselineDir), before, "refusal must not touch the baseline");
+});
+
+test("--update-baseline copies current over baseline when the refresh env var is set", () => {
+  const tmpRoot = fs.mkdtempSync(path.join(os.tmpdir(), "davinci-bench-compare-"));
+  try {
+    fs.cpSync(fixtureDir, tmpRoot, { recursive: true });
+    const result = runCompare(
+      [
+        "--budgets",
+        "budgets.toml",
+        "--baseline",
+        "baseline",
+        "--results",
+        "within-tolerance/current",
+        "--update-baseline",
+      ],
+      { cwd: tmpRoot, refreshEnv: "1" },
+    );
+    assert.equal(result.stderr, "");
+    assert.equal(result.status, 0, result.stdout);
+    assert.equal(
+      result.stdout,
+      "bench-compare: budgets=budgets.toml baseline=baseline results=within-tolerance/current\n" +
+        "updated baseline fixture_parse_small\n" +
+        "updated baseline fixture_transform_medium\n" +
+        "bench-compare: baseline updated (2 benches) under baseline\n",
+    );
+    assert.deepEqual(
+      baselineSnapshot(path.join(tmpRoot, "baseline")),
+      baselineSnapshot(path.join(tmpRoot, "within-tolerance", "current")),
+      "the refreshed baseline must be a byte copy of the current reports",
+    );
+  } finally {
+    fs.rmSync(tmpRoot, { recursive: true, force: true });
+  }
+});
