+++ tools/davinci/lib/bench-verdicts.mjs
@@ -1,0 +1,111 @@
+// Verdict rows for the davinci bench budget gate
+// (tools/davinci/bench-compare.mjs): one row per bench id, in sorted order,
+// so the stdout is a stable, diffable oracle.
+//
+// Gates, per registered bench
+//   wall p50 — against the BASELINE report (never the budget number), with
+//     the entry's tolerance applied in exact integer arithmetic (BigInt, the
+//     tolerance scaled to basis points).
+//   allocs   — deterministic, so gated exactly: any change fails, including
+//     a change to or from null.
+//   rss      — report-only (see the [bench] field docs in budgets.toml).
+//
+// Registry drift is a failure in both directions — silent drift is the enemy:
+// a budgets entry with no current result is a disappeared bench, and a result
+// with no entry is an unregistered (therefore ungated) bench.
+
+export function byBenchId(a, b) {
+  return a < b ? -1 : a > b ? 1 : 0;
+}
+
+function formatCount(value) {
+  return value === null ? "n/a" : String(value);
+}
+
+function formatRss(value) {
+  return value === null ? "n/a" : `${value}B`;
+}
+
+function wallLimit(baselineP50, toleranceBp) {
+  return (BigInt(baselineP50) * (10000n + BigInt(toleranceBp))) / 10000n;
+}
+
+export function compare(budgets, baselineReports, currentReports) {
+  const rows = [];
+  let breaches = 0;
+  let gatedOk = 0;
+  let reportOnly = 0;
+  const ids = [
+    ...new Set([...budgets.keys(), ...currentReports.keys(), ...baselineReports.keys()]),
+  ].sort(byBenchId);
+  for (const id of ids) {
+    const budget = budgets.get(id);
+    const current = currentReports.get(id);
+    const baseline = baselineReports.get(id);
+    if (budget == null) {
+      const where = current != null ? "current" : "baseline";
+      rows.push(
+        `FAIL ${id} unregistered bench (${where} result has no budgets.toml [bench] entry)`,
+      );
+      breaches += 1;
+      continue;
+    }
+    if (current == null) {
+      rows.push(`FAIL ${id} bench disappeared (budgets.toml entry has no current result)`);
+      breaches += 1;
+      continue;
+    }
+    const stats =
+      `wall_p50 ${current.wall_ns.p50}ns allocs ${formatCount(current.allocs)} ` +
+      `rss ${formatRss(current.rss_peak_bytes)}`;
+    if (baseline == null) {
+      rows.push(`unbaselined ${id} ${stats} (report-only)`);
+      reportOnly += 1;
+      continue;
+    }
+    if (budget.wall_p50_ns === 0) {
+      rows.push(`unrecorded ${id} ${stats} (report-only: budgets.toml baseline not yet recorded)`);
+      reportOnly += 1;
+      continue;
+    }
+    const limit = wallLimit(baseline.wall_ns.p50, budget.toleranceBp);
+    const benchRows = [];
+    if (BigInt(current.wall_ns.p50) > limit) {
+      benchRows.push(
+        `FAIL ${id} wall_p50 ${current.wall_ns.p50}ns > limit ${limit}ns ` +
+          `(baseline ${baseline.wall_ns.p50}ns + ${budget.toleranceBp / 100}% tolerance)`,
+      );
+    }
+    if (current.allocs !== baseline.allocs) {
+      benchRows.push(
+        `FAIL ${id} allocs ${formatCount(baseline.allocs)} -> ${formatCount(current.allocs)} ` +
+          `(exact gate: allocs are deterministic)`,
+      );
+    }
+    if (benchRows.length > 0) {
+      rows.push(...benchRows);
+      breaches += benchRows.length;
+    } else {
+      rows.push(
+        `ok ${id} wall_p50 ${current.wall_ns.p50}ns ` +
+          `(baseline ${baseline.wall_ns.p50}ns limit ${limit}ns) ` +
+          `allocs ${formatCount(current.allocs)} rss ${formatRss(current.rss_peak_bytes)}`,
+      );
+      gatedOk += 1;
+    }
+  }
+  return { rows, breaches, gatedOk, reportOnly };
+}
+
+export function reconciliationOnly(budgets, currentReports) {
+  const rows = [];
+  const ids = [...new Set([...budgets.keys(), ...currentReports.keys()])].sort(byBenchId);
+  for (const id of ids) {
+    if (!budgets.has(id)) {
+      rows.push(`FAIL ${id} unregistered bench (current result has no budgets.toml [bench] entry)`);
+    } else if (!currentReports.has(id)) {
+      rows.push(`FAIL ${id} bench disappeared (budgets.toml entry has no current result)`);
+    }
+  }
+  return rows;
+}
